+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Issue #663: Fixed tool Buffer Analysis where layer visibility updates would not be recognized in custom trees. Now, all visible layers at tool starting time are available within the tool.
 - Issue #778: Fixed layer information display selection in metadata window that now always has names options in its select input.
 - Issue #788: Legend not updating correctly
+- Download print Buttons will now be blue at first and grey once the button was clicked.
 
 ---